requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.2.5
djangorestframework==3.16.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
psycopg2-binary
pytest
pytest-django
pytest-cov
ruff

